Implement get active models api call

+++ app/src/main/java/ru/kima/intelligentchat/presentation/connection/overview/events/COUiEvent.kt
@@ -7,6 +7,6 @@ sealed interface COUiEvent {
     sealed interface COSnackbar {
         data object NoUser : COSnackbar
         data object ErrorGetKudos : COSnackbar
-        data class ShowKudos(val kudos: Float) : COSnackbar
+        data class ShowKudos(val kudos: Double) : COSnackbar
     }
 }
