+++ domain/src/main/java/ru/kima/intelligentchat/domain/horde/model/UserInfo.kt
@@ -3,5 +3,5 @@ package ru.kima.intelligentchat.domain.horde.model
 data class UserInfo(
     val userName: String = String(),
     val id: Int = 0,
-    val kudos: Float = 0.0f
+    val kudos: Double = 0.0
 )
